Add store management

+++ webapp/src/app/manager/user-management/user-management.component.ts
@@ -1,8 +1,11 @@
-import {Component, OnInit} from '@angular/core';
-import {UserManagementService} from "./user-management.service";
-import {IUser} from "../../core/auth/user.service";
-import {UserModalService} from "../../service/user-modal.service";
-import {NotificationService} from "../../layouts/notification/notification.service";
+import { MDBModalRef } from 'ng-uikit-pro-standard';
+import { ConfirmModalComponent } from './../../modal/confirm-modal/confirm-modal.component';
+import { MDBModalService } from 'ng-uikit-pro-standard';
+import { Component, OnInit } from '@angular/core';
+import { UserManagementService } from './user-management.service';
+import { IUser } from '../../core/auth/user.service';
+import { UserModalService } from '../../service/user-modal.service';
+import { NotificationService } from '../../layouts/notification/notification.service';
 
 @Component({
   selector: 'app-user-management',
@@ -11,40 +14,52 @@ import {NotificationService} from "../../layouts/notification/notification.servi
 })
 export class UserManagementComponent implements OnInit {
   users: IUser[] = [];
+  confirmModalRef: MDBModalRef;
 
   constructor(
     private userService: UserManagementService,
     private userModalService: UserModalService,
-    private nofiService: NotificationService) {
-  }
+    private modalService: MDBModalService,
+    private notiService: NotificationService
+  ) {}
 
   ngOnInit(): void {
     this.fetchUsers();
   }
 
   fetchUsers(): void {
-    this.userService.fetchAll().subscribe(users =>
-      this.users = users
-    );
+    this.userService.fetchAll().subscribe((users) => (this.users = users));
   }
 
   deleteUser(id: number): void {
-    this.userService.delete(id).subscribe(() => {
-      this.users = this.users.filter(user => user.id !== id);
-      this.nofiService.showSuccess('Delete successfully!');
+    this.confirmModalRef = this.modalService.show(ConfirmModalComponent, {
+      containerClass: 'modal fade top',
+      class: 'modal-dialog modal-frame modal-top',
+      data: {
+        key: 'DeleteStaff',
+      },
+    });
+
+    this.confirmModalRef.content.action.subscribe(({value, key}) => {
+      if (key === "DeleteStaff" && value === ConfirmModalComponent.YES) {
+        this.userService.delete(id).subscribe(() => {
+          this.users = this.users.filter((user) => user.id !== id);
+          this.notiService.showSuccess('Delete successfully!');
+        });
+        this.confirmModalRef.hide();
+      }
     })
   }
 
   showDetailsModal(id: number): void {
-    this.userService.fetchById(id).subscribe(user => {
+    this.userService.fetchById(id).subscribe((user) => {
       this.userModalService.showDetailModel(user);
-    })
+    });
   }
 
   showUpdateModal(id: number): void {
-    this.userService.fetchById(id).subscribe(user => {
+    this.userService.fetchById(id).subscribe((user) => {
       this.userModalService.showUpdateModal(user);
-    })
+    });
   }
-
 }
